Extract CSharp field declarator partial
